Fix bug in rnazWindow.pl.
If a short alignment is encountered the window size variable
is adjusted. But it is never resetted, so if I have a short
alignment block of 50 any alignment blocks below were
sliced with a window size of 50. If we had a very short alignment
block alignments suddenly disappeared because they fell below
the limit. Thanks to Zhang Li / Peking University for reporting this bug.